pipeline/requirements.txt
fal-client>=0.8.1
openai>=2.6.1
python-dotenv>=1.0.0
